+++ eds-aic-20230930/src/Models/CreateAICloudPhoneRequest.php
@@ -1,0 +1,188 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Edsaic\V20230930\Models;
+
+use AlibabaCloud\Dara\Model;
+
+class CreateAICloudPhoneRequest extends Model
+{
+    /**
+     * @var int
+     */
+    public $amount;
+
+    /**
+     * @var bool
+     */
+    public $autoPay;
+
+    /**
+     * @var string
+     */
+    public $bandwidthPackageId;
+
+    /**
+     * @var string
+     */
+    public $bizRegionId;
+
+    /**
+     * @var string
+     */
+    public $imageId;
+
+    /**
+     * @var string
+     */
+    public $instanceGroupName;
+
+    /**
+     * @var string
+     */
+    public $instanceGroupSpec;
+
+    /**
+     * @var int
+     */
+    public $period;
+
+    /**
+     * @var string
+     */
+    public $periodUnit;
+
+    /**
+     * @var string
+     */
+    public $policyGroupId;
+
+    /**
+     * @var string
+     */
+    public $promotionId;
+    protected $_name = [
+        'amount' => 'Amount',
+        'autoPay' => 'AutoPay',
+        'bandwidthPackageId' => 'BandwidthPackageId',
+        'bizRegionId' => 'BizRegionId',
+        'imageId' => 'ImageId',
+        'instanceGroupName' => 'InstanceGroupName',
+        'instanceGroupSpec' => 'InstanceGroupSpec',
+        'period' => 'Period',
+        'periodUnit' => 'PeriodUnit',
+        'policyGroupId' => 'PolicyGroupId',
+        'promotionId' => 'PromotionId',
+    ];
+
+    public function validate()
+    {
+        parent::validate();
+    }
+
+    public function toArray($noStream = false)
+    {
+        $res = [];
+        if (null !== $this->amount) {
+            $res['Amount'] = $this->amount;
+        }
+
+        if (null !== $this->autoPay) {
+            $res['AutoPay'] = $this->autoPay;
+        }
+
+        if (null !== $this->bandwidthPackageId) {
+            $res['BandwidthPackageId'] = $this->bandwidthPackageId;
+        }
+
+        if (null !== $this->bizRegionId) {
+            $res['BizRegionId'] = $this->bizRegionId;
+        }
+
+        if (null !== $this->imageId) {
+            $res['ImageId'] = $this->imageId;
+        }
+
+        if (null !== $this->instanceGroupName) {
+            $res['InstanceGroupName'] = $this->instanceGroupName;
+        }
+
+        if (null !== $this->instanceGroupSpec) {
+            $res['InstanceGroupSpec'] = $this->instanceGroupSpec;
+        }
+
+        if (null !== $this->period) {
+            $res['Period'] = $this->period;
+        }
+
+        if (null !== $this->periodUnit) {
+            $res['PeriodUnit'] = $this->periodUnit;
+        }
+
+        if (null !== $this->policyGroupId) {
+            $res['PolicyGroupId'] = $this->policyGroupId;
+        }
+
+        if (null !== $this->promotionId) {
+            $res['PromotionId'] = $this->promotionId;
+        }
+
+        return $res;
+    }
+
+    public function toMap($noStream = false)
+    {
+        return $this->toArray($noStream);
+    }
+
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['Amount'])) {
+            $model->amount = $map['Amount'];
+        }
+
+        if (isset($map['AutoPay'])) {
+            $model->autoPay = $map['AutoPay'];
+        }
+
+        if (isset($map['BandwidthPackageId'])) {
+            $model->bandwidthPackageId = $map['BandwidthPackageId'];
+        }
+
+        if (isset($map['BizRegionId'])) {
+            $model->bizRegionId = $map['BizRegionId'];
+        }
+
+        if (isset($map['ImageId'])) {
+            $model->imageId = $map['ImageId'];
+        }
+
+        if (isset($map['InstanceGroupName'])) {
+            $model->instanceGroupName = $map['InstanceGroupName'];
+        }
+
+        if (isset($map['InstanceGroupSpec'])) {
+            $model->instanceGroupSpec = $map['InstanceGroupSpec'];
+        }
+
+        if (isset($map['Period'])) {
+            $model->period = $map['Period'];
+        }
+
+        if (isset($map['PeriodUnit'])) {
+            $model->periodUnit = $map['PeriodUnit'];
+        }
+
+        if (isset($map['PolicyGroupId'])) {
+            $model->policyGroupId = $map['PolicyGroupId'];
+        }
+
+        if (isset($map['PromotionId'])) {
+            $model->promotionId = $map['PromotionId'];
+        }
+
+        return $model;
+    }
+}